services/auth/requirements.txt
bcrypt==4.0.1
# Core Dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
aiokafka==0.9.0
kafka-python==2.0.2
minio==7.1.17
httpx==0.25.1
email-validator==1.3.1

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1
